+++ src/main/java/lotto/controller/LottoGenerator.java
@@ -1,4 +1,7 @@
-package lotto.domain;
+package lotto.controller;
+
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
 
 import java.util.ArrayList;
 import java.util.Collections;
